requirements.txt
amqp==1.4.9
anyjson==0.3.3
beautifulsoup4==4.8.1
billiard==3.3.0.23
bs4==0.0.1
celery==3.1.26.post2
certifi==2019.9.11
chardet==3.0.4
dj-database-url==0.5.0
Django==2.0.13
gunicorn==20.0.0
idna==2.8
kombu==3.0.37
pytz==2019.3
requests==2.22.0
soupsieve==1.9.4
sqlparse==0.3.0
urllib3==1.25.6
whitenoise==4.1.4
psycopg2==2.6.2